check whther there is an old goods number before saving plan

+++ protected/bll/Plan.php
@@ -4,20 +4,28 @@ class Plan extends PlanModel
 {
     public static function createPlan($data)
     {
-        $plan = new Plan();
-        $plan->goods_number = $data['goods_number'];
-        $plan->client_id = $data['client_id'];
-        $plan->deadline_time = strtotime($data['deadline']);
-        $plan->create_time = time();
-        if($plan->save()){
-            $productIdList = Product::createProduct($data['productData'], $plan);
-            if($productIdList){
-                $partIdList = ProductPart::createProductPart($data['partData'], $productIdList, $plan);
-                if(!$partIdList){
-                    Product::deleteByIdList($productIdList);
+        $condition = "goods_number = :goods_number";
+        $params = array(
+            ":goods_number" => $data['goods_number'],
+        );
+        if(self::model()->find($condition, $params)){
+            return 0;
+        }else{
+            $plan = new Plan();
+            $plan->goods_number = $data['goods_number'];
+            $plan->client_id = $data['client_id'];
+            $plan->deadline_time = strtotime($data['deadline']);
+            $plan->create_time = time();
+            if($plan->save()){
+                $productIdList = Product::createProduct($data['productData'], $plan);
+                if($productIdList){
+                    $partIdList = ProductPart::createProductPart($data['partData'], $productIdList, $plan);
+                    if(!$partIdList){
+                        Product::deleteByIdList($productIdList);
+                    }
+                }else{
+                    $plan->delete();
                 }
-            }else{
-                $plan->delete();
             }
         }
     }
